# docs/openapi.yaml
openapi: "3.1.0"
info:
  title: FlowLink API
  version: "1.0.0"
  description: >-
    FlowLink compliance, identity, and invoicing API for autonomous agent
    payments. Provides sanctions screening, KYA credential issuance, invoice
    lifecycle management, webhook delivery, and platform analytics.
  contact:
    name: FlowLink
    url: https://flowlink.io
  license:
    name: MIT
servers:
  - url: "{baseUrl}/v1"
    description: FlowLink API v1
    variables:
      baseUrl:
        default: "http://localhost:3001"
        description: Base URL of the API server
security:
  - ApiKeyHeader: []
  - BearerAuth: []
...
components:
  securitySchemes:
    ApiKeyHeader:
      type: apiKey
      in: header
      name: X-API-Key
      description: API key passed via X-API-Key header
    BearerAuth:
      type: http
      scheme: bearer
      description: API key or JWT passed via Authorization Bearer header
  schemas:
    ErrorResponse:
      type: object
      required: [success, error]
      properties:
        success:
          type: boolean
          const: false
        error:
          type: object
          required: [code, message]
          properties:
            code:
              type: string
              description: Machine-readable error code
            message:
              type: string
              description: Human-readable error message
            details:
              description: Validation error details (array of { path, message })
...
  responses:
...
    Unauthorized:
      description: Missing or invalid authentication
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
...
paths:
...
  /analytics/volume:
    get:
      tags: [Analytics]
      summary: Transaction volume over time
      operationId: analyticsVolume
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        "200":
          description: Volume data bucketed by time period
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/compliance:
    get:
      tags: [Analytics]
      summary: Compliance decision breakdown
      operationId: analyticsCompliance
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        "200":
          description: Compliance breakdown with percentages
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/risk:
    get:
      tags: [Analytics]
      summary: Risk score distribution
      operationId: analyticsRisk
      parameters:
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
        - name: granularity
          in: query
          schema:
            type: string
            enum: [hour, day, week, month]
            default: day
      responses:
        "200":
          description: Risk distribution data with summary stats
        "401":
          $ref: "#/components/responses/Unauthorized"
  /analytics/agents:
    get:
      tags: [Analytics]
      summary: Top agents by volume
      operationId: analyticsAgents
      parameters:
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
            minimum: 1
            maximum: 100
        - name: from
          in: query
          schema:
            type: string
            format: date-time
        - name: to
          in: query
          schema:
            type: string
            format: date-time
      responses:
        "200":
          description: Top agents data
        "401":
          $ref: "#/components/responses/Unauthorized"